requirements.txt
numpy
librosa
tensorflow
opencv-python
ultralytics
paho-mqtt
requests
